# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/dental-lab/cases": {
      "get": {
        "operationId": "DentalLabCasesController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DentalLabCaseStatus"
            }
          },
          {
            "name": "tagId",
            "required": false,
            "in": "query",
            "description": "Lab-side only: filter to a tag.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "default": 200,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-cases"]
      }
    },
    "/api/dental-lab/cases/{id}": {
      "get": {
        "operationId": "DentalLabCasesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-cases"]
      },
      "patch": {
        "operationId": "DentalLabCasesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLabCaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-cases"]
      }
    },
    "/api/dental-lab/cases/{id}/transitions": {
      "post": {
        "operationId": "DentalLabCasesController_transition",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionLabCaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-cases"]
      }
    },
    "/api/dental-lab/cases/{id}/files/presign": {
      "post": {
        "operationId": "DentalLabCasesController_presignFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignLabCaseFileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-cases"]
      }
    },
    "/api/dental-lab/cases/{id}/files/{fileId}/confirm": {
      "post": {
        "operationId": "DentalLabCasesController_confirmFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-cases"]
      }
    },
    "/api/dental-lab/cases/{id}/files/{fileId}": {
      "delete": {
        "operationId": "DentalLabCasesController_deleteFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-cases"]
      }
    },
    "/api/dental-lab/cases/{id}/phases": {
      "get": {
        "operationId": "DentalLabCasesController_listPhases",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-cases"]
      }
    },
    "/api/dental-lab/cases/{id}/phases/advance": {
      "post": {
        "operationId": "DentalLabCasesController_advancePhase",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvancePhaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-cases"]
      }
    },
    "/api/dental-lab/cases/{id}/notes": {
      "get": {
        "operationId": "DentalLabCasesController_listNotes",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-cases"]
      },
      "post": {
        "operationId": "DentalLabCasesController_createNote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-cases"]
      }
    },
    "/api/dental-lab/cases/{id}/notes/{noteId}": {
      "patch": {
        "operationId": "DentalLabCasesController_updateNote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "noteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-cases"]
      },
      "delete": {
        "operationId": "DentalLabCasesController_deleteNote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "noteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-cases"]
      }
    },
    "/api/dental-lab/cases/{id}/messages": {
      "get": {
        "operationId": "DentalLabCasesController_listMessages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-cases"]
      },
      "post": {
        "operationId": "DentalLabCasesController_createMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-cases"]
      }
    },
    "/api/dental-lab/cases/{id}/messages/{messageId}": {
      "delete": {
        "operationId": "DentalLabCasesController_deleteMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-cases"]
      }
    },
    "/api/dental-lab/cases/{id}/shipment": {
      "get": {
        "operationId": "DentalLabCasesController_getShipment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-cases"]
      },
      "post": {
        "operationId": "DentalLabCasesController_upsertShipment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertShipmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-cases"]
      }
    },
...
    "/api/clinic/lab-cases/{id}": {
      "get": {
        "operationId": "ClinicLabCasesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["clinic-lab-cases"]
      },
      "patch": {
        "operationId": "ClinicLabCasesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLabCaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["clinic-lab-cases"]
      }
    },
    "/api/clinic/lab-cases/{id}/transitions": {
      "post": {
        "operationId": "ClinicLabCasesController_transition",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionLabCaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["clinic-lab-cases"]
      }
    },
    "/api/clinic/lab-cases/{id}/files/presign": {
      "post": {
        "operationId": "ClinicLabCasesController_presignFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignLabCaseFileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["clinic-lab-cases"]
      }
    },
    "/api/clinic/lab-cases/{id}/files/{fileId}/confirm": {
      "post": {
        "operationId": "ClinicLabCasesController_confirmFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["clinic-lab-cases"]
      }
    },
    "/api/clinic/lab-cases/{id}/files/{fileId}": {
      "delete": {
        "operationId": "ClinicLabCasesController_deleteFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["clinic-lab-cases"]
      }
    },
    "/api/clinic/lab-cases/{id}/phases": {
      "get": {
        "operationId": "ClinicLabCasesController_listPhases",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["clinic-lab-cases"]
      }
    },
    "/api/clinic/lab-cases/{id}/messages": {
      "get": {
        "operationId": "ClinicLabCasesController_listMessages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["clinic-lab-cases"]
      },
      "post": {
        "operationId": "ClinicLabCasesController_createMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["clinic-lab-cases"]
      }
    },
    "/api/clinic/lab-cases/{id}/messages/{messageId}": {
      "delete": {
        "operationId": "ClinicLabCasesController_deleteMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["clinic-lab-cases"]
      }
    },
    "/api/clinic/lab-cases/{id}/shipment": {
      "get": {
        "operationId": "ClinicLabCasesController_getShipment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["clinic-lab-cases"]
      }
    },
...
  },
  "info": {
    "title": "ClinIQ API",
    "description": "Multi-tenant clinic management API",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4005"
    }
  ],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DentalLabCaseStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "SUBMITTED",
          "IN_PROGRESS",
          "AWAITING_PICKUP",
          "SHIPPED",
          "DELIVERED",
          "CANCELLED",
          "REJECTED"
        ]
      },
      "DentalLabCaseUrgency": {
        "type": "string",
        "enum": ["STANDARD", "URGENT"]
      },
      "UpdateLabCaseDto": {
        "type": "object",
        "properties": {
          "urgency": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DentalLabCaseUrgency"
              }
            ]
          },
          "dueAt": {
            "type": "object"
          },
          "formData": {
            "type": "object"
          },
          "patientLabel": {
            "type": "object"
          },
          "doctorLabel": {
            "type": "object"
          },
          "deliveryCenter": {
            "type": "object"
          },
          "notes": {
            "type": "object"
          },
          "unitPrice": {
            "type": "object",
            "description": "Lab-only override of unit price (centavos). Used with ADJUST_ON_ORDER products."
          }
        }
      },
      "TransitionLabCaseDto": {
        "type": "object",
        "properties": {
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DentalLabCaseStatus"
              }
            ]
          },
          "reason": {
            "type": "string"
          }
        },
        "required": ["status"]
      },
      "PresignLabCaseFileDto": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string",
            "description": "Original filename — passed through to S3 metadata."
          },
          "mimeType": {
            "type": "string",
            "example": "application/octet-stream"
          },
          "sizeBytes": {
            "type": "number",
            "description": "File size in bytes."
          }
        },
        "required": ["filename", "mimeType", "sizeBytes"]
      },
      "AdvancePhaseDto": {
        "type": "object",
        "properties": {
          "phase": {
            "type": "string",
            "description": "Target phase (must match one entry from the product's phases array). Omit to advance to the next phase in declared order."
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "CreateNoteDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "description": "Internal note body (lab-only)."
          }
        },
        "required": ["body"]
      },
      "UpdateNoteDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          }
        },
        "required": ["body"]
      },
      "CreateMessageDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "description": "Chat message body."
          }
        },
        "required": ["body"]
      },
      "UpsertShipmentDto": {
        "type": "object",
        "properties": {
          "carrier": {
            "type": "object",
            "example": "LBC"
          },
          "trackingNumber": {
            "type": "object",
            "example": "AB1234567PH"
          },
          "notes": {
            "type": "object"
          }
        }
      },
...
    }
  }
}